+++ src/fastga_he/models/propulsion/components/source/pemfc/unit_tests/test_pemfc.py
@@ -18,8 +18,8 @@ from ..components.sizing_pemfc_drag import SizingPEMFCDrag
 
 from ..components.perf_fuel_consumption import PerformancesPEMFCFuelConsumption
 from ..components.perf_fuel_consumed import PerformancesPEMFCFuelConsumed
-from ..components.perf_layer_voltage import PerformancesSinglePEMFCVoltageSimple
-from ..components.perf_layer_voltage import PerformancesSinglePEMFCVoltageAnalytical
+from ..components.perf_pemfc_layer_voltage import PerformancesSinglePEMFCVoltageSimple
+from ..components.perf_pemfc_layer_voltage import PerformancesSinglePEMFCVoltageAnalytical
 from ..components.perf_pemfc_current_density import PerformancesCurrentDensity
 from ..components.perf_maximum_current import PerformancesMaximumCurrent
 from ..components.perf_maximum_power import PerformancesMaximumPower
@@ -27,8 +27,8 @@ from ..components.perf_pemfc_efficiency import PerformancesPEMFCEfficiency
 from ..components.perf_pemfc_power import PerformancesPEMFCPower
 from ..components.perf_pemfc_specific_power import PerformancesPEMFCSpecificPower
 from ..components.perf_pemfc_voltage import PerformancesPEMFCVoltage
-from ..components.perf_operating_pressure import PerformancesOperatingPressure
-from ..components.perf_operating_temperature import PerformancesOperatingTemperature
+from ..components.perf_pemfc_operating_pressure import PerformancesOperatingPressure
+from ..components.perf_pemfc_operating_temperature import PerformancesOperatingTemperature
 from ..components.perf_pemfc_expect_specific_power import (
     PerformancesPEMFCMaxSpecificPowerFuelCellSystem,
 )
@@ -41,7 +41,9 @@ from ..components.perf_pemfc_expect_power_density import (
 from ..components.perf_pemfc_expect_power_density import (
     PerformancesPEMFCMaxPowerDensityFuelCellStack,
 )
-from ..components.perf_analytical_voltage_adjustment import PerformancesAnalyticalVoltageAdjustment
+from ..components.perf_pemfc_analytical_voltage_adjustment import (
+    PerformancesAnalyticalVoltageAdjustment,
+)
 
 from ..components.cstr_ensure import ConstraintsEffectiveAreaEnsure
 from ..components.cstr_enforce import ConstraintsEffectiveAreaEnforce
